requirements.txt
python-binance==1.0.19
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.26.2
ccxt==4.2.25
ta==0.11.0
pyyaml==6.0.1
requests==2.31.0
pytest==7.4.3
matplotlib==3.8.2
tabulate==0.9.0
flask==3.0.0
flask-socketio==5.3.6
python-socketio==5.10.0
eventlet==0.34.2
